+++ tfjs-core/src/ops/ops.ts
@@ -76,6 +76,7 @@ export {maxPool} from './max_pool';
 export {maxPool3d} from './max_pool_3d';
 export {maxPoolWithArgmax} from './max_pool_with_argmax';
 export {maximum} from './maximum';
+export {mean} from './mean';
 export {min} from './min';
 export {minimum} from './minimum';
 export {mod} from './mod';
