+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -65,6 +65,7 @@ const CHAINED_OPS = [
   'max',
   'maximum',
   'maxPool',
+  'mean',
   'min',
   'minimum',
   'mod',
