+++ tfjs-core/src/tests.ts
@@ -109,6 +109,7 @@ import './ops/max_pool_test';
 import './ops/max_pool_with_argmax_test';
 import './ops/max_test';
 import './ops/mean_squared_error_test';
+import './ops/mean_test';
 import './ops/min_test';
 import './ops/moments_test';
 import './ops/moving_average_test';
